Add Options::reset()
Add reset() to Options to reset the status of parsing command line
arguments. The need for this comes from changing the test framework
to Catch. getopt is used for command line parsing and it has some
global variables keeping the state of parsing process. If
arguments are parsed a second time, using the same object or even
another object, getopt variables are left in the state where they
were after the first parsing operation and second parsing will fail.
Using check, this doesn't matter because it forks for individual
tests. Catch doesn't. Now reset() is called in Options' destructor
or you can call the static reset() function itself.