Set defaults when configure is run
Set defaults where the config directory is, the names of default config
and database when configure is run. Create that directory and empty
config.
I should probably omit this and think how the program should behave on
different situations, such as there is no directory or can't open
config. Should the program create the directory and config, ignore
errors or fail on error.